backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.115.13",
    "ffmpeg>=1.4",
    "ffmpeg-python>=0.2.0",
    "google-cloud-aiplatform>=1.98.0",
    "google-generativeai>=0.8.5",
    "langchain-core>=0.3.66",
    "langgraph>=0.4.8",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "starlette>=0.46.2",
    "uvicorn>=0.34.3",
    "yt-dlp>=2025.6.9",
]
